+++ Scene.cpp
@@ -66,14 +66,14 @@ bool Scene::load(const std::string& filename) {
                     } else {
                         // Build from TRS
                         if (!node.scale.empty()) {
-                            transform[0] *= node.scale[0];
-                            transform[5] *= node.scale[1];
-                            transform[10] *= node.scale[2];
+                            transform[0] *= (float)node.scale[0];
+                            transform[5] *= (float)node.scale[1];
+                            transform[10] *= (float)node.scale[2];
                         }
                         
                         if (!node.rotation.empty()) {
-                            float qx = node.rotation[0], qy = node.rotation[1];
-                            float qz = node.rotation[2], qw = node.rotation[3];
+                            float qx = (float)node.rotation[0], qy = (float)node.rotation[1];
+                            float qz = (float)node.rotation[2], qw = (float)node.rotation[3];
                             
                             float m[9];
                             m[0] = 1 - 2*(qy*qy + qz*qz);
@@ -99,9 +99,9 @@ bool Scene::load(const std::string& filename) {
                         }
                         
                         if (!node.translation.empty()) {
-                            transform[12] = node.translation[0];
-                            transform[13] = node.translation[1];
-                            transform[14] = node.translation[2];
+                            transform[12] = (float)node.translation[0];
+                            transform[13] = (float)node.translation[1];
+                            transform[14] = (float)node.translation[2];
                         }
                     }
                     break;
